remove test duplication from build and publish task
